app/requirements.txt
boto3==1.24.91
Django>=4.1.2,<4.2.0
Djangorestframework>=3.14.0,<3.15.0
psycopg2>=2.9.4,<2.10.0
Pillow>9.2.0<9.3.0
# flake8>=3.6.0,<3.7.0
drf-yasg==1.21.4
django-storages==1.13.1
gunicorn==20.1.0
drf-social-oauth2==1.2.1
unidecode==1.3.2
pyjwt==2.1.0
django-admin-interface==0.22.1



# psycopg2-binary==2.8.3